Patch 1: March 9, 2015
At the thread-safe configuration (_FS_REENTRANT == 1), an unexpected
operation to the volume can result a wrong volume lock. For example,
f_open() ==> media change ==> f_close(), in this case the volume gets
left locked and following function calls to the volume will be locked
out until a forced initialization of the volume by f_mount().  To solve
this problem, apply this patch to the ff.c.